Create an app episode category based on the app name
