+++ opal/core/application.py
@@ -92,7 +92,7 @@ class OpalApplication(object):
             index=0
         )
     ]
-    default_episode_category = 'Inpatient'
+    default_episode_category = None
 
     @classmethod
     def get_core_javascripts(klass, namespace):
