+++ opal/core/scaffold.py
@@ -202,6 +202,7 @@ def start_project(name, USERLAND_HERE):
     9. Create a superuser
     10. Initialise our git repo
     11. Load referencedata shipped with Opal
+    12. Create the episode category template
     """
 
     project_dir = USERLAND_HERE/name
@@ -288,6 +289,12 @@ def start_project(name, USERLAND_HERE):
 
     # 11. Load referencedata shipped with Opal
     manage('load_lookup_lists')
+
+    # 12. Create app detail template
+    nix.mv(
+        app_dir/'templates/detail/app.html',
+        app_dir/'templates/detail/{0}.html'.format(name)
+    )
 
 
 def _strip_non_user_fields(schema):
