+++ opal/scaffolding/scaffold/app/templates/detail/app.html
