+++ opal/tests/test_core_episodes.py
@@ -7,6 +7,7 @@ from django.contrib.auth.models import User
 
 from opal.core import test
 from opal.models import UserProfile, Patient
+from opal.tests import episode_categories
 
 from opal.core import episodes
 
@@ -21,38 +22,38 @@ class EpisodeCategoryTestCase(test.OpalTestCase):
         )
         self.patient = Patient.objects.create()
         self.inpatient_episode = self.patient.create_episode(
-            category_name=episodes.InpatientEpisode.display_name
+            category_name=episode_categories.InpatientEpisode.display_name
         )
 
     def test_episode_categories(self):
         self.assertIn(
-            episodes.InpatientEpisode, episodes.EpisodeCategory.list()
+            episode_categories.InpatientEpisode, episodes.EpisodeCategory.list()
         )
 
     def test_visible_to(self):
         self.assertTrue(
-            episodes.InpatientEpisode.episode_visible_to(
+            episode_categories.InpatientEpisode.episode_visible_to(
                 self.inpatient_episode, self.user)
         )
 
     def test_visible_to_restricted_only(self):
         self.assertFalse(
-            episodes.InpatientEpisode.episode_visible_to(
+            episode_categories.InpatientEpisode.episode_visible_to(
                 self.inpatient_episode,
                 self.restricted_user)
         )
 
     def test_for_category(self):
-        self.assertEqual(episodes.InpatientEpisode,
+        self.assertEqual(episode_categories.InpatientEpisode,
                          episodes.EpisodeCategory.get('inpatient'))
 
     def test_is_active(self):
-        category = episodes.InpatientEpisode(self.inpatient_episode)
+        category = episode_categories.InpatientEpisode(self.inpatient_episode)
         self.assertTrue(category.is_active())
 
     def test_is_active_end_date_set(self):
         self.inpatient_episode.end = datetime.date.today()
-        category = episodes.InpatientEpisode(self.inpatient_episode)
+        category = episode_categories.InpatientEpisode(self.inpatient_episode)
         self.assertFalse(category.is_active())
 
     def test_get_stages(self):
@@ -61,7 +62,7 @@ class EpisodeCategoryTestCase(test.OpalTestCase):
             'Followup',
             'Discharged'
         ]
-        category = episodes.InpatientEpisode(self.inpatient_episode)
+        category = episode_categories.InpatientEpisode(self.inpatient_episode)
         self.assertEqual(stages, category.get_stages())
 
     def test_get_stages_unchanged(self):
@@ -70,21 +71,21 @@ class EpisodeCategoryTestCase(test.OpalTestCase):
             'Followup',
             'Discharged'
         ]
-        category = episodes.InpatientEpisode(self.inpatient_episode)
+        category = episode_categories.InpatientEpisode(self.inpatient_episode)
         stages1 = category.get_stages()
         stages1.append('hahhahahaha')
         self.assertEqual(stages, category.get_stages())
 
     def test_has_stage_true(self):
-        category = episodes.InpatientEpisode(self.inpatient_episode)
+        category = episode_categories.InpatientEpisode(self.inpatient_episode)
         self.assertTrue(category.has_stage('Inpatient'))
 
     def test_has_stage_case_sensitive(self):
-        category = episodes.InpatientEpisode(self.inpatient_episode)
+        category = episode_categories.InpatientEpisode(self.inpatient_episode)
         self.assertFalse(category.has_stage('inpatient'))
 
     def test_has_stage_false(self):
-        category = episodes.InpatientEpisode(self.inpatient_episode)
+        category = episode_categories.InpatientEpisode(self.inpatient_episode)
         self.assertFalse(category.has_stage('Airegin'))
 
     def test_can_monkeypatch_stages(self):
